+++ docs/options/options.js
@@ -6,7 +6,7 @@
   };
 
   // <define:process.env>
-  var define_process_env_default = { BUILD_TIME: "2023-01-27T00:31:19.470Z", VERSION: "0.2.26", PROD: "1", IMMERSIVE_TRANSLATE_INJECTED_CSS: `.immersive-translate-target-translation-pre-whitespace {
+  var define_process_env_default = { BUILD_TIME: "2023-01-27T15:18:29.175Z", VERSION: "0.2.27", PROD: "1", IMMERSIVE_TRANSLATE_INJECTED_CSS: `.immersive-translate-target-translation-pre-whitespace {
   white-space: pre-wrap !important;
 }
 
@@ -7807,7 +7807,9 @@ body {
         selectors: [
           ".markdown-title",
           ".markdown-body",
-          ".Layout-sidebar p"
+          ".Layout-sidebar p",
+          "div > span.search-match",
+          "#responsive-meta-container p"
         ],
         excludeSelectors: [
           ".css-truncate",
@@ -9295,6 +9297,26 @@ body {
       lastTextNode = treeWalker.currentNode;
     return lastTextNode ? lastTextNode.parentElement : null;
   }
+  function getRealUrl() {
+    if (!getIsInIframe())
+      return globalThis.location.href;
+    try {
+      let currentUrl = globalThis.location.href;
+      if (new URL(currentUrl).protocol === "about:") {
+        if (globalThis.location.ancestorOrigins && globalThis.location.ancestorOrigins.length > 0)
+          return globalThis.location.ancestorOrigins[0];
+        let href = "";
+        try {
+          href = globalThis.parent.location.href;
+        } catch {
+        }
+        return href || (globalThis.location != globalThis.parent.location ? document.referrer : document.location.href);
+      } else
+        return currentUrl;
+    } catch {
+    }
+    return globalThis.location.href;
+  }
 
   // dom/mark_containers.ts
   function markContainers(containers, rule) {
@@ -9810,7 +9832,11 @@ body {
         languageByLocal: currentLanguageByLocal,
         languageByClient: currentPageLanguageByClient2
       };
-      if (!isSameTargetLanguage(result, targetLanguage)) {
+      if (paragraphEntities.set(newParagraph.id, {
+        ...newParagraph,
+        state: "Original",
+        observers: []
+      }), !isSameTargetLanguage(result, targetLanguage)) {
         if (excludeLanguages.length > 0 && excludeLanguages.some((language) => isSameTargetLanguage(result, language)))
           return;
         filterdParagraphs.push(newParagraph);
@@ -10657,7 +10683,7 @@ body {
   }
 
   // dom/translate_page.ts
-  var pageStatus = "Original", currentParagraphIds = [], waitToTranslateParagraphIds = /* @__PURE__ */ new Set(), allNewDynamicElements = [], allIntersectionObserver = [], allResizebleObserver = [], currentNewDynamicElements = [], oldUrl = globalThis.location.href.split("#")[0], currentTranslatedTextLength = 0, globalContext, debounceTranslateCurrentQueue = debounce(translateCurrentQueue, 300), debounceTranslateNewDynamicNodes = debounce(
+  var pageStatus = "Original", currentParagraphIds = [], waitToTranslateParagraphIds = /* @__PURE__ */ new Set(), allNewDynamicElements = [], allIntersectionObserver = [], allResizebleObserver = [], currentNewDynamicElements = [], oldUrl = getRealUrl().split("#")[0], currentTranslatedTextLength = 0, globalContext, debounceTranslateCurrentQueue = debounce(translateCurrentQueue, 300), debounceTranslateNewDynamicNodes = debounce(
     translateNewDynamicNodes,
     200
   ), env3 = getEnv(), isProd2 = env3.PROD === "1", mutationObserver, titleMutationObserver, originalPageTitle = "";
@@ -10796,7 +10822,7 @@ body {
     if (pageStatus === "Translating")
       return;
     setPageTranslatedStatus("Translating");
-    let ctx = await getGlobalContext(globalThis.location.href);
+    let ctx = await getGlobalContext(getRealUrl());
     if (!ctx.state.isAutoTranslate && ctx.config.tempTranslateDomainMinutes > 0) {
       let now = Date.now(), currentDomain = new URL(ctx.url).hostname, currentTempTranslationDomains = ctx.localConfig.tempTranslationUrlMatches || [], index = currentTempTranslationDomains.findIndex(
         (item) => item.match === currentDomain && item.expiredAt > now
@@ -10850,19 +10876,19 @@ body {
     return globalContext;
   }
   async function toggleTranslateTheMainPage() {
-    getPageStatus() === "Original" ? await translateTheMainPage() : (getPageStatus() === "Translated" || getPageStatus() === "Error") && (globalContext = await getGlobalContext(globalThis.location.href), globalContext.state.translationArea !== "main" ? await translateTheMainPage() : restorePage());
+    getPageStatus() === "Original" ? await translateTheMainPage() : (getPageStatus() === "Translated" || getPageStatus() === "Error") && (globalContext = await getGlobalContext(getRealUrl()), globalContext.state.translationArea !== "main" ? await translateTheMainPage() : restorePage());
   }
   async function translateTheMainPage() {
-    globalContext = await getGlobalContext(globalThis.location.href), globalContext.state.translationArea = "main", await translatePage();
+    globalContext = await getGlobalContext(getRealUrl()), globalContext.state.translationArea = "main", await translatePage();
   }
   async function translateTheWholePage() {
-    globalContext = await getGlobalContext(globalThis.location.href), globalContext.state.translationArea = "body", await translatePage();
+    globalContext = await getGlobalContext(getRealUrl()), globalContext.state.translationArea = "body", await translatePage();
   }
   async function toggleTranslateTheWholePage() {
-    getPageStatus() === "Original" ? await translateTheWholePage() : (getPageStatus() === "Translated" || getPageStatus() === "Error") && (globalContext = await getGlobalContext(globalThis.location.href), globalContext.state.translationArea !== "body" ? (globalContext.state.translationArea = "body", await translatePage()) : restorePage());
+    getPageStatus() === "Original" ? await translateTheWholePage() : (getPageStatus() === "Translated" || getPageStatus() === "Error") && (globalContext = await getGlobalContext(getRealUrl()), globalContext.state.translationArea !== "body" ? (globalContext.state.translationArea = "body", await translatePage()) : restorePage());
   }
   async function translateToThePageEndImmediately() {
-    globalContext = await getGlobalContext(globalThis.location.href), globalContext.state.translationArea = "body", globalContext.state.translationStartMode = "immediate", await translatePage(), await translateNewDynamicNodes(globalContext);
+    globalContext = await getGlobalContext(getRealUrl()), globalContext.state.translationArea = "body", globalContext.state.translationStartMode = "immediate", await translatePage(), await translateNewDynamicNodes(globalContext);
   }
   async function translateTitle(ctx) {
     let pageTitle = document.title;
@@ -11016,7 +11042,7 @@ body {
     let inlineAndIgnoreAndTextTags = rule.inlineTags.concat(rule.excludeTags).concat("#text", "BR");
     mutationObserver = new MutationObserver(function(mutations) {
       mutations.forEach((mutation) => {
-        let currentUrl = globalThis.location.href;
+        let currentUrl = getRealUrl();
         if (currentUrl.split("#")[0] !== oldUrl && rule.observeUrlChange) {
           oldUrl = currentUrl.split("#")[0], clean(), disableMutatinObserver(), setTimeout(() => {
             log_default.debug("url changed, reinit page"), initPage();
@@ -11067,7 +11093,7 @@ body {
     return !1;
   }
   async function initPage() {
-    let isInIframe = getIsInIframe(), ctx = await getGlobalContext(globalThis.location.href);
+    let isInIframe = getIsInIframe(), ctx = await getGlobalContext(getRealUrl());
     ctx.rule.urlChangeDelay && await delay(ctx.rule.urlChangeDelay);
     let lang = ctx.sourceLanguage;
     lang === "auto" ? (isMonkey() ? lang = await detectLanguage({
@@ -13540,9 +13566,7 @@ body {
       return {
         ...payload
       };
-    let { config, translationService } = ctx, generalConfig = config.translationGeneralConfig, services = config.translationServices, defaultTranslationEngine = translationService, serviceConfig = services[defaultTranslationEngine] || {};
-    payload.sentences = payload.sentences.map((sentence) => sentence);
-    let noCacheSentences = [], finalResult = {
+    let { config, translationService } = ctx, generalConfig = config.translationGeneralConfig, services = config.translationServices, defaultTranslationEngine = translationService, serviceConfig = services[defaultTranslationEngine] || {}, noCacheSentences = [], finalResult = {
       sentences: Array(payload.sentences.length)
     }, sourceLength = payload.sentences.length, sentenceIndex = -1;
     if (config.cache)
@@ -15119,11 +15143,14 @@ body {
       e3.preventDefault();
       let currentSettings = { ...settings };
       delete currentSettings.rules, handleChangeValue(currentSettings);
-    }, defaultUserRulePlaceholder = {
-      matches: "www.google.com",
-      selectors: [".title", '[data-testid="title"]'],
-      excludeSelectors: ["footer"]
-    };
+    }, defaultUserRulePlaceholder = `[{
+  "matches": "www.google.com",
+  "selectors": [".title"]
+}, {
+  "matches": "*.twitter.com",
+  "selectors": [".text"],
+  "excludeSelectors": ["footer"]
+}]`;
     return config ? /* @__PURE__ */ p5("div", {
       children: [
         /* @__PURE__ */ p5("div", {
@@ -15160,7 +15187,7 @@ body {
             /* @__PURE__ */ p5("textarea", {
               rows: 10,
               onChange: handleChangeRules,
-              placeholder: JSON.stringify(defaultUserRulePlaceholder, null, 2),
+              placeholder: defaultUserRulePlaceholder,
               value: settings.rules ? JSON.stringify(settings.rules || [], null, 2) : ""
             }),
             /* @__PURE__ */ p5("div", {